Remove colors module, add custom widget
urwid uses widget wrapping to build custom widgets. WordWidget is one
such example. The class will be moved to a separate module when more
features are slowly added.

All colors have also been replaced with the defined palette. I believe
palette should also be a separate module when customization and theming
features are added in the future.

For now, the unicode strings are formatted with u"%s". I'm not sure if
there is a f-strings or a better alternative for improved readability.