Add end screen, quit hotkey and reset test hotkey
An input filter is added to the main loop to distinguish between hotkey bindings and typing test.

A working reset feature has been implemented by instantiating a new
instance of the mode class that was passed as an argument. Although it
works, I feel like there is a better way of resetting the event loop
instead of recreating a new widget every time.

The end screen has been added.